# openapi-schema.yaml
openapi: 3.0.0
info:
  title: Tarkov Casino API
  version: 1.0.0
  description: |
    API for Tarkov Casino gaming platform
    
    Note: This schema is used by Cloudflare API Shield for request validation.
    The production server is protected by Cloudflare with schema validation,
    rate limiting, and security headers.
servers:
  - url: https://tarkov.juanis.cool/api
    description: Production server (protected by Cloudflare API Shield)
...
components:
  securitySchemes:
    AppwriteSession:
      type: apiKey
      in: header
      name: X-Appwrite-Session
      description: Appwrite session secret
    AppwriteUser:
      type: apiKey
      in: header
      name: X-Appwrite-User-Id
      description: Appwrite user ID
...
paths:
...
  /user/history:
    get:
      tags:
        - user
      summary: Get game history
      security:
        - AppwriteSession: []
        - AppwriteUser: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: game_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Game history
        '401':
          description: Authentication required
...
  /statistics/basic:
    get:
      tags:
        - statistics
      summary: Get basic statistics
      security:
        - AppwriteSession: []
        - AppwriteUser: []
      responses:
        '200':
          description: Basic statistics
        '401':
          description: Authentication required
  /statistics/advanced:
    get:
      tags:
        - statistics
      summary: Get advanced statistics
      security:
        - AppwriteSession: []
        - AppwriteUser: []
      parameters:
        - name: gameType
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Advanced statistics
        '401':
          description: Authentication required